requirements.txt
pandas
matplotlib
telebot
psycopg2
PILLOW
python-dotenv